@@ -1,66 +1,49 @@
 {
-  "name": "recomposition",
-  "version": "0.1.0",
+  "name": "recompositionnew",
+  "version": "0.0.0",
   "license": "MIT",
-  "angular-cli": {},
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build",
+    "build": "ng build --prod",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
-  "keywords": [],
-  "author": "Skyvory",
   "dependencies": {
-    "@angular/animations": "^4.0.2",
-    "@angular/common": "^4.0.2",
-    "@angular/compiler": "^4.0.2",
-    "@angular/compiler-cli": "^4.0.2",
-    "@angular/core": "^4.0.2",
-    "@angular/flex-layout": "^2.0.0-beta.7",
-    "@angular/forms": "^4.0.2",
-    "@angular/http": "^4.0.2",
-    "@angular/material": "^2.0.0-beta.3",
-    "@angular/platform-browser": "^4.0.2",
-    "@angular/platform-browser-dynamic": "^4.0.2",
-    "@angular/platform-server": "^4.0.2",
-    "@angular/router": "^4.0.2",
-    "angular2-jwt": "^0.1.18",
-    "bootstrap": "^3.3.7",
+    "@angular/animations": "^5.0.0",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.0.0",
+    "@angular/http": "^5.0.0",
+    "@angular/platform-browser": "^5.0.0",
+    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@angular/router": "^5.0.0",
     "core-js": "^2.4.1",
-    "moment": "^2.16.0",
-    "ng2-file-upload": "^1.2.0",
-    "ng2-pagination": "^1.0.1",
-    "ngx-clipboard": "^5.0.9",
-    "reflect-metadata": "^0.1.8",
-    "rxjs": "^5.1.0",
-    "three": "^0.85.2",
-    "typescript": "^2.2.2",
-    "zone.js": "^0.8.4"
+    "rxjs": "^5.5.2",
+    "zone.js": "^0.8.14"
   },
   "devDependencies": {
-    "@angular/cli": "^1.0.4",
-    "@angular/compiler-cli": "^4.0.0",
-    "@types/jasmine": "2.5.38",
+    "@angular/cli": "1.6.3",
+    "@angular/compiler-cli": "^5.0.0",
+    "@angular/language-service": "^5.0.0",
+    "@types/jasmine": "~2.5.53",
+    "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
-    "codelyzer": "~2.0.0",
-    "concurrently": "^3.0.0",
-    "es6-promise-loader": "^1.0.2",
-    "jasmine-core": "~2.5.2",
-    "jasmine-spec-reporter": "~3.2.0",
-    "karma": "~1.4.1",
-    "karma-chrome-launcher": "~2.0.0",
+    "codelyzer": "^4.0.1",
+    "jasmine-core": "~2.6.2",
+    "jasmine-spec-reporter": "~4.1.0",
+    "karma": "~1.7.0",
+    "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
+    "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.0",
-    "ts-node": "~2.0.0",
-    "tslint": "~4.5.0",
-    "typescript": "~2.1.0",
-    "webdriver-manager": "10.2.5"
+    "protractor": "~5.1.2",
+    "ts-node": "~3.2.0",
+    "tslint": "~5.7.0",
+    "typescript": "~2.4.2"
   }
 }
