@@ -1,14 +1,14 @@
-import { RecompositionPage } from './app.po';
+import { AppPage } from './app.po';
 
-describe('recomposition App', function() {
-  let page: RecompositionPage;
+describe('recompositionnew App', () => {
+  let page: AppPage;
 
   beforeEach(() => {
-    page = new RecompositionPage();
+    page = new AppPage();
   });
 
-  it('should display message saying app works', () => {
+  it('should display welcome message', () => {
     page.navigateTo();
-    expect(page.getParagraphText()).toEqual('app works!');
+    expect(page.getParagraphText()).toEqual('Welcome to app!');
   });
 });
