@@ -1,17 +1,18 @@
 import { Injectable } from '@angular/core';
-import { Http, Headers, Response, RequestOptions } from '@angular/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { Headers, Response, RequestOptions } from '@angular/http';
 import { Observable } from 'rxjs';
 import 'rxjs/add/operator/map';
 import { contentHeaders } from './common/headers';
-// import 'rxjs/add/operator/toPromise';
 import { Constant } from './const.config';
+// import { catchError, map, tap } from 'rxjs/operators';
 
 @Injectable()
 export class AuthenticationService {
 	public token: string;
 	public authorization_bearer: string;
 
-	constructor(private http: Http) {
+	constructor(private http: HttpClient) {
 		var currentUser = localStorage.getItem('recomposition_token');
 		this.token = currentUser;
 		this.authorization_bearer = 'Bearer ' + this.token;
@@ -22,13 +23,14 @@ export class AuthenticationService {
 
 	public option:any = this.constructOptions();
 	private constructOptions():any {
-		// let headers = new Headers({ 'Authorization': 'Bearer ' + this.token });
 		let token = localStorage.getItem('recomposition_token');
-		let headers = new Headers();
-		headers.append('Authorization', 'Bearer ' + token);
-		headers.append('Accept', 'application/json, text/plain');
-		headers.append('Content-Type', 'application/json;charset=utf-8');
-		let options = new RequestOptions({ headers: headers });
+		let headers = new HttpHeaders();
+		headers = headers.append('Authorization', 'Bearer ' + token);
+		headers = headers.append('Accept', 'application/json, text/plain');
+		headers = headers.append('Content-Type', 'application/json;charset=utf-8');
+		let options = {
+			headers: headers
+		};
 		return options;
 	}
 
@@ -37,42 +39,43 @@ export class AuthenticationService {
 	}
 	private constructParamOptions(params):any {
 		let token = localStorage.getItem('recomposition_token');
-		let headers = new Headers();
-		headers.append('Authorization', 'Bearer ' + token);
-		headers.append('Accept', 'application/json, text/plain');
-		headers.append('Content-Type', 'application/json;charset=utf-8');
-		let options = new RequestOptions({ headers: headers, search: params });
+		let headers = new HttpHeaders();
+		headers = headers.append('Authorization', 'Bearer ' + token);
+		headers = headers.append('Accept', 'application/json, text/plain');
+		headers = headers.append('Content-Type', 'application/json;charset=utf-8');
+		let options = {
+			headers: headers,
+			search: params
+		};
 		return options;
 	}
 
 	public optionOnUpload:any = this.constructOptionsOnUpload();
 	private constructOptionsOnUpload():any {
 		let token = localStorage.getItem('recomposition_token');
-		let headers = new Headers();
-		headers.append('Authorization', 'Bearer ' + token);
-		headers.append('Accept', 'application/json, text/plain');
-		headers.append('Content-Type', 'multipart/form-data; charset=utf-8; boundary=' + Math.random().toString().substr(2));
-		let options = new RequestOptions({ headers: headers });
+		let headers = new HttpHeaders();
+		headers = headers.append('Authorization', 'Bearer ' + token);
+		headers = headers.append('Accept', 'application/json, text/plain');
+		headers = headers.append('Content-Type', 'multipart/form-data; charset=utf-8; boundary=' + Math.random().toString().substr(2));
+		let options = {
+			headers: headers
+		};
 		return options;
 	}
 
 	login(username, password): Observable<boolean> {
-		// return this.http.post('http://localhost/record/public/api/authenticate', JSON.stringify({username:username, password:password}), {headers: this.headers})
-		// .toPromise()
-		// .then(response => {
-		// 	console.log(response);
-		// });
+		console.log(username, password);
 		let body = JSON.stringify({ username: username, password: password });
+		console.log(body);
 		return this.http.post(Constant.API_PATH + `authenticate`, body, {headers: contentHeaders })
-			.map((response: Response) => {
-				console.log(response);
-				let token = response.json() && response.json().token;
+			.map((response:any) => {
+				let token = response && response.token;
 				if(token) {
 					this.token = token;
 
 					localStorage.setItem('recomposition_token', token);
 
-					let user = response.json().user;
+					let user = response.user;
 					localStorage.setItem('recomposition_user', JSON.stringify(user));
 					this.user = user;
 
@@ -81,14 +84,16 @@ export class AuthenticationService {
 					return false;
 				}
 			});
+		
 	}
 
 	isJwtTokenValid():Observable<boolean> {
 		let token = localStorage.getItem('recomposition_token');
+		console.log(token);
 		return this.http.get(Constant.API_PATH + 'authentication/validity', this.option)
 			.map(
-				(response:Response) => {
-					if(response.json().user) {
+				(response:any) => {
+					if(response.user) {
 						this.user = JSON.parse(localStorage.getItem('recomposition_user'));
 						console.log(this.user);
 						return true;
@@ -99,8 +104,8 @@ export class AuthenticationService {
 				}
 			)
 			.catch(error => {
-				if(error.json().error) {
-					return Observable.throw(error.json().error);
+				if(error.error) {
+					return Observable.throw(error.error);
 				}
 			});
 	}
